+++ ansible_collections/scruffaluff/bootware/roles/desktop_environment/files/gnome.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 # Show weekday in top bar date.
 gsettings set org.gnome.desktop.interface clock-show-weekday true