+++ install.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 #
 # Install Bootware for FreeBSD, MacOS, or Linux systems.
 
@@ -6,9 +6,8 @@
 #
 # Flags:
 #   -e: Exit immediately when a command pipeline fails.
-#   -o: Persist nonzero exit codes through a Bash pipe.
 #   -u: Throw an error when an unset variable is encountered.
-set -eou pipefail
+set -eu
 
 #######################################
 # Show CLI help information.
@@ -43,7 +42,7 @@ assert_cmd() {
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v "${1}")" ]]; then
+  if [ ! -x "$(command -v "${1}")" ]; then
     error "Cannot find required ${1} command on computer."
   fi
 }
@@ -58,7 +57,7 @@ assert_cmd() {
 #   Parent directory of Bootware script.
 #######################################
 configure_shell() {
-  local export_cmd="export PATH=\"${1}:\$PATH\"" profile shell_name
+  export_cmd="export PATH=\"${1}:\$PATH\""
   shell_name="$(basename "${SHELL}")"
 
   case "${shell_name}" in
@@ -76,7 +75,8 @@ configure_shell() {
       profile="${HOME}/.config/fish/config.fish"
       ;;
     *)
-      error "Shell ${shell_name} is not supported."
+      echo "Shell ${shell_name} is not supported for configuration."
+      return 0
       ;;
   esac
 
@@ -90,7 +90,7 @@ configure_shell() {
 #   Writes error message to stderr.
 #######################################
 error() {
-  local bold_red='\033[1;31m' default='\033[0m'
+  bold_red='\033[1;31m' default='\033[0m'
   printf "${bold_red}error${default}: %s\n" "${1}" >&2
   exit 1
 }
@@ -101,36 +101,86 @@ error() {
 #   Writes error message to stderr.
 #######################################
 error_usage() {
-  local bold_red='\033[1;31m' default='\033[0m'
+  bold_red='\033[1;31m' default='\033[0m'
   printf "${bold_red}error${default}: %s\n" "${1}" >&2
   printf "Run 'bootware --help' for usage.\n" >&2
   exit 2
 }
 
 #######################################
+# Find command to elevate as super user.
+#######################################
+find_super() {
+  # Flags:
+  #   -v: Only show file path of command.
+  #   -x: Check if file exists and execute permission is granted.
+  if [ -x "$(command -v sudo)" ]; then
+    echo 'sudo'
+  elif [ -x "$(command -v doas)" ]; then
+    echo 'doas'
+  else
+    echo ''
+  fi
+}
+
+#######################################
+# Install Bash shell.
+# Arguments:
+#   Super user command for installation.
+#######################################
+install_bash() {
+  # Do not quote the outer super parameter expansion. Shell will error due to be
+  # being unable to find the "" command.
+  if [ -x "$(command -v apk)" ]; then
+    ${1:+"${1}"} apk update
+    ${1:+"${1}"} apk add bash
+  elif [ -x "$(command -v apt-get)" ]; then
+    ${1:+"${1}"} apt-get update
+    ${1:+"${1}"} apt-get install --quiet --yes bash
+  elif [ -x "$(command -v dnf)" ]; then
+    ${1:+"${1}"} dnf check-update || {
+      code="$?"
+      [ "${code}" -ne 100 ] && exit "${code}"
+    }
+    ${1:+"${1}"} dnf install --assumeyes bash
+  elif [ -x "$(command -v pacman)" ]; then
+    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${1:+"${1}"} pacman --noconfirm --sync bash
+  elif [ -x "$(command -v pkg)" ]; then
+    ${1:+"${1}"} pkg update
+    ${1:+"${1}"} pkg install --yes bash
+  elif [ -x "$(command -v zypper)" ]; then
+    ${1:+"${1}"} zypper update --no-confirm
+    ${1:+"${1}"} zypper install --no-confirm bash
+  else
+    error 'No supported package manager found to install Bash.'
+  fi
+}
+
+#######################################
 # Install completion scripts for Bootware.
 # Arguments:
-#   Whether to use sudo for installation.
+#   Super user elevation command.
 #   Whether to install for entire system.
 #   GitHub version reference.
 #######################################
 install_completions() {
-  local repo_url="https://raw.githubusercontent.com/scruffaluff/bootware/${3}"
-  local bash_url="${repo_url}/completions/bootware.bash"
-  local fish_url="${repo_url}/completions/bootware.fish"
+  repo_url="https://raw.githubusercontent.com/scruffaluff/bootware/${3}"
+  bash_url="${repo_url}/completions/bootware.bash"
+  fish_url="${repo_url}/completions/bootware.fish"
 
   # Flags:
   #   -z: Check if the string has zero length or is null.
-  if [[ -z "${2:-}" ]]; then
+  if [ -z "${2:-}" ]; then
     # Do not use long form --parents flag for mkdir. It is not supported on
     # MacOS.
-    ${1:+sudo} mkdir -p '/etc/bash_completion.d'
-    ${1:+sudo} curl -LSfs "${bash_url}" -o '/etc/bash_completion.d/bootware.bash'
-    ${1:+sudo} chmod 664 '/etc/bash_completion.d/bootware.bash'
+    ${1:+"${1}"} mkdir -p '/etc/bash_completion.d'
+    ${1:+"${1}"} curl -LSfs "${bash_url}" -o '/etc/bash_completion.d/bootware.bash'
+    ${1:+"${1}"} chmod 664 '/etc/bash_completion.d/bootware.bash'
 
-    ${1:+sudo} mkdir -p '/etc/fish/completions'
-    ${1:+sudo} curl -LSfs "${fish_url}" -o '/etc/fish/completions/bootware.fish'
-    ${1:+sudo} chmod 664 '/etc/fish/completions/bootware.fish'
+    ${1:+"${1}"} mkdir -p '/etc/fish/completions'
+    ${1:+"${1}"} curl -LSfs "${fish_url}" -o '/etc/fish/completions/bootware.fish'
+    ${1:+"${1}"} chmod 664 '/etc/fish/completions/bootware.fish'
   else
     mkdir -p "${HOME}/.config/fish/completions"
     curl -LSfs "${fish_url}" -o "${HOME}/.config/fish/completions/bootware.fish"
@@ -141,16 +191,16 @@ install_completions() {
 #######################################
 # Install Man pages for Bootware.
 # Arguments:
-#   Whether to use sudo for installation.
+#   Super user command for installation.
 #   GitHub version reference.
 #######################################
 install_man() {
-  local man_url="https://raw.githubusercontent.com/scruffaluff/bootware/${2}/bootware.1"
+  man_url="https://raw.githubusercontent.com/scruffaluff/bootware/${2}/bootware.1"
 
   # Do not use long form --parents flag for mkdir. It is not supported on MacOS.
-  ${1:+sudo} mkdir -p '/usr/local/share/man/man1'
-  ${1:+sudo} curl -LSfs "${man_url}" -o '/usr/local/share/man/man1/bootware.1'
-  ${1:+sudo} chmod 664 '/usr/local/share/man/man1/bootware.1'
+  ${1:+"${1}"} mkdir -p '/usr/local/share/man/man1'
+  ${1:+"${1}"} curl -LSfs "${man_url}" -o '/usr/local/share/man/man1/bootware.1'
+  ${1:+"${1}"} chmod 664 '/usr/local/share/man/man1/bootware.1'
 }
 
 #######################################
@@ -165,7 +215,7 @@ log() {
   #
   # Flags:
   #   -z: Check if string has zero length.
-  if [[ -z "${BOOTWARE_NOLOG:-}" ]]; then
+  if [ -z "${BOOTWARE_NOLOG:-}" ]; then
     echo "$@"
   fi
 }
@@ -174,11 +224,12 @@ log() {
 # Script entrypoint.
 #######################################
 main() {
-  local dst_dir dst_file='/usr/local/bin/bootware' src_url use_sudo='' \
-    user_install version='main'
+  dst_file='/usr/local/bin/bootware'
+  super=''
+  version='main'
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       --debug)
         set -o xtrace
@@ -208,53 +259,63 @@ main() {
   done
 
   assert_cmd curl
-
   src_url="https://raw.githubusercontent.com/scruffaluff/bootware/${version}/bootware.sh"
 
-  # Use sudo for system installation if user did not give the --user, does not
-  # own the file, and is not root.
+  # Use super user command for system installation if user did not give the
+  # --user, does not own the file, and is not root. Do not use long form --user
+  # flag for id command. It is not supported on MacOS.
   #
   # Flags:
+  #   -u: Print only the user id.
   #   -w: Check if file exists and is writable.
   #   -z: Check if the string has zero length or is null.
-  if [[ -z "${user_install:-}" && ! -w "${dst_file}" && "${EUID}" -ne 0 ]]; then
-    assert_cmd sudo
-    use_sudo='true'
+  if [ -z "${user_install:-}" ] && [ ! -w "${dst_file}" ] &&
+    [ "$(id -u)" -ne 0 ]; then
+    super="$(find_super)"
+    if [ -z "${super:-}" ]; then
+      error 'No command found to elevate user.'
+    fi
   fi
-  dst_dir="$(dirname "${dst_file}")"
 
+  # Install Bash shell if necessary.
+  #
+  # Flags:
+  #   -v: Only show file path of command.
+  if [ ! -x "$(command -v bash)" ]; then
+    install_bash "${super}"
+  fi
+
+  dst_dir="$(dirname "${dst_file}")"
   log 'Installing Bootware...'
 
-  # Do not quote the sudo parameter expansion. Bash will error due to be being
-  # unable to find the "" command.
-  # Do not use long form --parents flag for mkdir. It is not supported on MacOS.
-  ${use_sudo:+sudo} mkdir -p "${dst_dir}"
-  ${use_sudo:+sudo} curl -LSfs "${src_url}" --output "${dst_file}"
-  ${use_sudo:+sudo} chmod 755 "${dst_file}"
+  # Do not quote the outer super parameter expansion. Shell will error due to be
+  # being unable to find the "" command. Do not use long form --parents flag for
+  # mkdir. It is not supported on MacOS.
+  ${super:+"${super}"} mkdir -p "${dst_dir}"
+  ${super:+"${super}"} curl -LSfs "${src_url}" --output "${dst_file}"
+  ${super:+"${super}"} chmod 755 "${dst_file}"
 
   # Add Bootware to shell profile if not in system path.
   #
   # Flags:
   #   -e: Check if file exists.
   #   -v: Only show file path of command.
-  if [[ ! -e "$(command -v bootware)" ]]; then
+  if [ ! -e "$(command -v bootware)" ]; then
     configure_shell "${dst_dir}"
     export PATH="${dst_dir}:${PATH}"
   fi
 
-  install_completions "${use_sudo}" "${user_install:-}" "${version}"
+  install_completions "${super}" "${user_install:-}" "${version}"
 
   # Installl man pages if a system install.
   #
   # Flags:
   #   -z: Check if the string has zero length or is null.
-  if [[ -z "${user_install:-}" ]]; then
-    install_man "${use_sudo}" "${version}"
+  if [ -z "${user_install:-}" ]; then
+    install_man "${super}" "${version}"
   fi
 
   log "Installed $(bootware --version)."
 }
 
-# Variable BASH_SOURCE cannot be used to load script as a library. Piping the
-# script to Bash gives the same BASH_SOURCE result as sourcing the script.
 main "$@"